requirements.txt
google-cloud-bigquery
google-cloud-storage
google-generativeai
